requirements.txt
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
black==21.12b0
boto3==1.17.58
botocore==1.20.112
cached-property==1.5.2
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
colorclass==2.2.0
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.10
django-allauth==0.44.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-jsignature==0.11
django-sendgrid-v5==0.9.1
django-storages==1.11.1
django-tenants==3.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
docx-mailmerge==0.5.0
et-xmlfile==1.1.0
eventlet==0.31.0
fdfgen==0.16.1
flake8==3.9.2
flower==1.0.0
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
httpie==2.4.0
humanize==3.13.1
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
jdcal==1.4.1
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.5
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
oauthlib==3.1.1
openpyxl==3.0.7
ordered-set==4.0.2
packaging==20.9
pathspec==0.9.0
Pillow==8.3.2
pip-upgrader==1.4.15
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.16
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.3
PySocks==1.7.1
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.1
python-http-client==3.3.2
python-memcached==1.59
python-telegram-bot==13.6
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
reportlab==3.6.7
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.4.2
sendgrid==6.7.1
sendgrid-python==0.1.1
sentry-sdk==1.5.4
simplejson==3.17.2
six==1.16.0
smtpapi==0.4.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.1
tenant-schemas-celery==1.0.1
terminaltables==3.1.0
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.4
webencodings==0.5.1
whitenoise==5.2.0
XlsxWriter==1.4.3
xlwt==1.3.0
zipp==3.4.1
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
black==21.12b0
boto3==1.17.58
botocore==1.20.112
cached-property==1.5.2
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
colorclass==2.2.0
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.10
django-allauth==0.44.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-jsignature==0.11
django-sendgrid-v5==0.9.1
django-storages==1.11.1
django-tenant-schemas==1.10.0
django-tenants==3.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
docx-mailmerge==0.5.0
et-xmlfile==1.1.0
eventlet==0.31.0
fdfgen==0.16.1
flake8==3.9.2
flower==1.0.0
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
httpie==2.4.0
humanize==3.13.1
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
jdcal==1.4.1
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.5
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
oauthlib==3.1.1
openpyxl==3.0.7
ordered-set==4.0.2
packaging==20.9
pathspec==0.9.0
Pillow==8.3.2
pip-upgrader==1.4.15
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.16
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.3
PySocks==1.7.1
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.1
python-http-client==3.3.2
python-memcached==1.59
python-telegram-bot==13.6
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
reportlab==3.6.7
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.4.2
sendgrid==6.7.1
sendgrid-python==0.1.1
sentry-sdk==1.5.4
simplejson==3.17.2
six==1.16.0
smtpapi==0.4.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.1
tenant-schemas-celery==1.0.1
terminaltables==3.1.0
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
typing-extensions==3.10.0.0
tzdata==2022.1
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.4
webencodings==0.5.1
whitenoise==5.2.0
XlsxWriter==1.4.3
xlwt==1.3.0
zipp==3.4.1
amqp==5.1.1
APScheduler==3.6.3
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
black==21.12b0
boto3==1.17.58
botocore==1.20.112
cached-property==1.5.2
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
colorclass==2.2.0
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.10
django-allauth==0.44.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-environ==0.4.5
django-filter==2.4.0
django-jsignature==0.11
django-sendgrid-v5==0.9.1
django-storages==1.11.1
django-tenant-schemas==1.10.0
django-tenants==3.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
docx-mailmerge==0.5.0
et-xmlfile==1.1.0
eventlet==0.31.0
fdfgen==0.16.1
flake8==3.9.2
flower==1.0.0
future==0.18.2
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
httpie==2.4.0
humanize==3.13.1
idna==2.10
importlib-metadata==4.5.0
iniconfig==1.1.1
jdcal==1.4.1
jmespath==0.10.0
kombu==5.1.0
lxml==4.6.5
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
oauthlib==3.1.1
openpyxl==3.0.7
ordered-set==4.0.2
packaging==20.9
pathspec==0.9.0
Pillow==8.3.2
pip-upgrader==1.4.15
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.16
psycopg2-binary==2.9.3
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.3
PySocks==1.7.1
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.1
python-http-client==3.3.2
python-memcached==1.59
python-telegram-bot==13.6
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
redis==3.5.3
reportlab==3.6.7
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
s3transfer==0.4.2
sendgrid==6.7.1
sendgrid-python==0.1.1
sentry-sdk==1.5.4
simplejson==3.17.2
six==1.16.0
smtpapi==0.4.7
sqlparse==0.4.2
starkbank-ecdsa==2.0.1
tenant-schemas-celery==1.0.1
terminaltables==3.1.0
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.3
tornado==6.1
typing-extensions==3.10.0.0
tzdata==2022.1
tzlocal==2.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
WeasyPrint==52.4
webencodings==0.5.1
whitenoise==5.2.0
XlsxWriter==1.4.3
xlwt==1.3.0
zipp==3.4.1
